disabling button fix and changing to Input component and moving key functions to hook
