+++ components/wallet/lib/index.ts
@@ -1,0 +1,1 @@
+export { useDepositTransaction } from './hooks';
